+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - added type wrapper for `EntityInfo` to allow for easier entity definition syntax in bindings
+- added source line reporting for `CommandOutput` type wrapper
 
 ## [3.4.0] - 2026-03-17